wagner_coach/fitness-backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.11.9
pydantic-settings==2.11.0

# Database
supabase==2.20.0

# AI/ML
openai==1.109.1

# Background Jobs
celery[redis]==5.5.3
redis==5.2.1

# Authentication
python-jose[cryptography]==3.5.0
python-multipart==0.0.6

# External Integrations
garminconnect==0.2.30

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.45.1

# Utilities
python-dotenv==1.1.1